requirements.txt
flask==3.1.1
python-dotenv==1.1.1
requests==2.32.3
PyMuPDF==1.26.1
pdfminer.six==20221105
spacy==3.8.7
scikit-learn==1.6.1
pandas==2.2.3
openpyxl==3.1.5
joblib==1.4.2
gunicorn==22.0.0
openai==0.28
tabulate 

